Replace `when` argument from add_validator with `condition` method
If you think of InputValidator as nodes in a graph, and add_validator()
as edges, then `when` was akin to adding conditions to edges. On further
reflection, conditions really belong on nodes; this will make a specific
feature feasible to implement in the future, which is, enabling real-time
feedback display on an InputValidator while its parent InputValidator is
disabled. With the conditions-on-edges approach, I couldn't figure out
how to implement partially-enabled forms without being mutually exclusive
with conditions.